Upload issue attachment fixes
