+++ src/main/kotlin/io/easybreezy/project/model/issue/Attachment.kt
@@ -30,6 +30,8 @@ class Attachment private constructor(id: EntityID<UUID>) : AggregateRoot<UUID>(i
                 this.issue = EntityID(issue, Attachments)
             }
         }
+
+        const val BUCKET = "issue-attachment"
     }
 
     fun add(attachments: List<File>) {
